+++ tests/commands/test_list.py
@@ -3,23 +3,12 @@
 
 import json
 
-
-from click.testing import CliRunner
-
 from platformio.commands.list import cli
 
-runner = CliRunner()
 
-
-def validate_output(result):
-    assert result.exit_code == 0
-    assert not result.exception
-    assert "error" not in result.output.lower()
-
-
-def test_list_json_output():
-    result = runner.invoke(cli, ["--json-output"])
-    validate_output(result)
+def test_list_json_output(clirunner, validate_cliresult):
+    result = clirunner.invoke(cli, ["--json-output"])
+    validate_cliresult(result)
     list_result = json.loads(result.output)
     assert isinstance(list_result, list)
     assert len(list_result)
@@ -27,7 +16,7 @@ def test_list_json_output():
     assert "titiva" in platforms
 
 
-def test_list_raw_output():
-    result = runner.invoke(cli)
-    validate_output(result)
+def test_list_raw_output(clirunner, validate_cliresult):
+    result = clirunner.invoke(cli)
+    validate_cliresult(result)
     assert "teensy" in result.output